Fix capitalization errors for act strings using leading $n anchors
In act strings, `$n` doesn't capitalize an actor's name, `$c` does.
`$c` should be used at the start of an act string, as most mobs will
have descriptions such as `a mob`, `the mob`, which should be
capitalized.

The MUD does something slightly annoying: it forces the first byte of
an act string to a capital. It is compensating for some old code,
perhaps. This behaviour allows the wrong anchor to be used everywhere.

Later, when colour escapes were added to start of act strings, the
capitalisation hack doesn't apply, and we don't capitalize the visible
text.

Fix-up some of these errors. Can't quite face changing it everywhere!