Don't keep server/player/ under version control: use a skeleton
It is annoying to have server/player/ under version control as local
testing often involves using players and making changes to the initial
files. Irrelevant or undesirable changes can often be made to the
"stock" distributed files.

Move this initial set of files into the separate server/player_skeleton/
directory, which is used to create the initial server/player/ structure.

Update documentation to discuss this.