Avoid crashing when recording bug information
In some situations (e.g. the player/ directory is not writable when
attempting to save a pfile) the server will report the wrong error and
then crash due to encountering another error during error handling.

This error relates to manipulating the "reserved file handle" fpReserve.
fpReserve may already be closed and we attempt to double-close it.

Rather than trying to cope with fights over fpReserve, just remove this
manipulation from the bug function. We are unlikely to ever hit OS
limits these days.