Log last command on memory allocation failures (#469)
Something wants to allocate a lot of memory: more than the slab
allocator supports. Suspect it is related to sending content to
player descriptors, given the size (128000 bytes).

Log the last player command before aborting in alloc_mem.
There is a much lower limit to how we can send to players, so it would
be good to figure out what is trying to work with the amount of memory.